requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
supabase==2.3.5
openai==1.12.0
flask-cors==4.0.0
python-jose==3.3.0 